+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+* Validate `FlushInterval` to prevent out-of-range values and
+  zero-interval spin loops.
+  ([#4860](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4860))
+
+* Fixed exceptions thrown from a user-supplied `Enrich`
+  callback preventing the `Activity` from being stopped.
+  ([#4900](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4900))
+
+* Fixed exceptions thrown by Redis profiler callbacks after
+  instrumentation is disposed.
+  ([#4905](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4905))
+
 ## 1.17.0-beta.1
 
 Released 2026-Jul-17